refactor Birdwatcher to use modules
