+++ src/utilities/index.js
@@ -1,0 +1,2 @@
+export { findUsernames } from "./findUsernames";
+export { applyTags } from "./applyTags";
